requirements.txt
pyserial
pyserial-asyncio
pycryptodome
tqdm[tui]
colorama[tui] ; platform_system == "Windows"
